requirements.txt
pymongo>=4.0
argon2-cffi>=21.3
PyNaCl>=1.5.0